fix: update biopath for vercel build to work; chore: rename post-actions.tsx to actions.tsx
